SAMOA-14: Move all streams under single package
