+++ samoa-api/src/main/java/org/apache/samoa/streams/ExampleStream.java
@@ -1,4 +1,4 @@
-package org.apache.samoa.moa.streams;
+package org.apache.samoa.streams;
 
 /*
  * #%L
