+++ samoa-api/src/main/java/org/apache/samoa/streams/StreamSourceProcessor.java
@@ -29,7 +29,6 @@ import org.apache.samoa.core.Processor;
 import org.apache.samoa.instances.Instance;
 import org.apache.samoa.instances.Instances;
 import org.apache.samoa.learners.InstanceContentEvent;
-import org.apache.samoa.moa.streams.InstanceStream;
 import org.apache.samoa.topology.Stream;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
