+++ samoa-api/src/main/java/org/apache/samoa/tasks/ClusteringEvaluation.java
@@ -26,10 +26,10 @@ import org.apache.samoa.evaluation.ClusteringEvaluatorProcessor;
 import org.apache.samoa.learners.Learner;
 import org.apache.samoa.learners.clusterers.simple.ClusteringDistributorProcessor;
 import org.apache.samoa.learners.clusterers.simple.DistributedClusterer;
-import org.apache.samoa.moa.streams.InstanceStream;
-import org.apache.samoa.moa.streams.clustering.ClusteringStream;
-import org.apache.samoa.moa.streams.clustering.RandomRBFGeneratorEvents;
 import org.apache.samoa.streams.ClusteringEntranceProcessor;
+import org.apache.samoa.streams.InstanceStream;
+import org.apache.samoa.streams.clustering.ClusteringStream;
+import org.apache.samoa.streams.clustering.RandomRBFGeneratorEvents;
 import org.apache.samoa.topology.ComponentFactory;
 import org.apache.samoa.topology.Stream;
 import org.apache.samoa.topology.Topology;
