+++ samoa-api/src/main/java/org/apache/samoa/tasks/PrequentialEvaluation.java
@@ -32,9 +32,9 @@ import org.apache.samoa.learners.ClassificationLearner;
 import org.apache.samoa.learners.Learner;
 import org.apache.samoa.learners.RegressionLearner;
 import org.apache.samoa.learners.classifiers.trees.VerticalHoeffdingTree;
-import org.apache.samoa.moa.streams.InstanceStream;
-import org.apache.samoa.moa.streams.generators.RandomTreeGenerator;
+import org.apache.samoa.streams.InstanceStream;
 import org.apache.samoa.streams.PrequentialSourceProcessor;
+import org.apache.samoa.streams.generators.RandomTreeGenerator;
 import org.apache.samoa.topology.ComponentFactory;
 import org.apache.samoa.topology.Stream;
 import org.apache.samoa.topology.Topology;
